+++ Sourcery/CodeGenerated/Description.generated.swift
@@ -1,12 +1,22 @@
-// Generated using Sourcery 0.5.1 — https://github.com/krzysztofzablocki/Sourcery
+// Generated using Sourcery 0.5.3 — https://github.com/krzysztofzablocki/Sourcery
 // DO NOT EDIT
+
+extension AssociatedValue {
+    override var description: String {
+        var string = "\(type(of: self)): "
+        string += "localName = \(self.localName), "
+        string += "externalName = \(self.externalName), "
+        string += "typeName = \(self.typeName)"
+        return string
+    }
+}
 
 extension Enum {
     override var description: String {
         var string = super.description
-        string += "cases = \(cases), "
-        string += "rawTypeName = \(rawTypeName), "
-        string += "hasAssociatedValues = \(hasAssociatedValues)"
+        string += "cases = \(self.cases), "
+        string += "rawTypeName = \(self.rawTypeName), "
+        string += "hasAssociatedValues = \(self.hasAssociatedValues)"
         return string
     }
 }
@@ -14,21 +24,26 @@ extension Enum {
 extension EnumCase {
     override var description: String {
         var string = "\(type(of: self)): "
-        string += "name = \(name), "
-        string += "rawValue = \(rawValue), "
-        string += "associatedValues = \(associatedValues), "
-        string += "annotations = \(annotations), "
-        string += "hasAssociatedValue = \(hasAssociatedValue)"
+        string += "name = \(self.name), "
+        string += "rawValue = \(self.rawValue), "
+        string += "associatedValues = \(self.associatedValues), "
+        string += "annotations = \(self.annotations), "
+        string += "hasAssociatedValue = \(self.hasAssociatedValue)"
         return string
     }
 }
 
-extension AssociatedValue {
+extension GenerationContext {
     override var description: String {
         var string = "\(type(of: self)): "
-        string += "localName = \(localName), "
-        string += "externalName = \(externalName), "
-        string += "typeName = \(typeName)"
+        string += "types = \(self.types), "
+        string += "typeByName = \(self.typeByName), "
+        string += "arguments = \(self.arguments), "
+        string += "classes = \(self.classes), "
+        string += "all = \(self.all), "
+        string += "protocols = \(self.protocols), "
+        string += "structs = \(self.structs), "
+        string += "enums = \(self.enums)"
         return string
     }
 }
@@ -36,18 +51,18 @@ extension AssociatedValue {
 extension Method {
     override var description: String {
         var string = "\(type(of: self)): "
-        string += "selectorName = \(selectorName), "
-        string += "parameters = \(parameters), "
-        string += "shortName = \(shortName), "
-        string += "returnTypeName = \(returnTypeName), "
-        string += "throws = \(`throws`), "
-        string += "accessLevel = \(accessLevel), "
-        string += "isStatic = \(isStatic), "
-        string += "isClass = \(isClass), "
-        string += "isInitializer = \(isInitializer), "
-        string += "isFailableInitializer = \(isFailableInitializer), "
-        string += "annotations = \(annotations), "
-        string += "attributes = \(attributes)"
+        string += "selectorName = \(self.selectorName), "
+        string += "parameters = \(self.parameters), "
+        string += "shortName = \(self.shortName), "
+        string += "returnTypeName = \(self.returnTypeName), "
+        string += "throws = \(self.throws), "
+        string += "accessLevel = \(self.accessLevel), "
+        string += "isStatic = \(self.isStatic), "
+        string += "isClass = \(self.isClass), "
+        string += "isInitializer = \(self.isInitializer), "
+        string += "isFailableInitializer = \(self.isFailableInitializer), "
+        string += "annotations = \(self.annotations), "
+        string += "attributes = \(self.attributes)"
         return string
     }
 }
@@ -55,20 +70,10 @@ extension Method {
 extension MethodParameter {
     override var description: String {
         var string = "\(type(of: self)): "
-        string += "argumentLabel = \(argumentLabel), "
-        string += "name = \(name), "
-        string += "typeName = \(typeName), "
-        string += "type = \(type), "
-        string += "typeAttributes = \(typeAttributes)"
-        return string
-    }
-}
-
-extension TupleType {
-    override var description: String {
-        var string = "\(type(of: self)): "
-        string += "name = \(name), "
-        string += "elements = \(elements)"
+        string += "argumentLabel = \(self.argumentLabel), "
+        string += "name = \(self.name), "
+        string += "typeName = \(self.typeName), "
+        string += "typeAttributes = \(self.typeAttributes)"
         return string
     }
 }
@@ -76,8 +81,17 @@ extension TupleType {
 extension TupleElement {
     override var description: String {
         var string = "\(type(of: self)): "
-        string += "name = \(name), "
-        string += "typeName = \(typeName)"
+        string += "name = \(self.name), "
+        string += "typeName = \(self.typeName)"
+        return string
+    }
+}
+
+extension TupleType {
+    override var description: String {
+        var string = "\(type(of: self)): "
+        string += "name = \(self.name), "
+        string += "elements = \(self.elements)"
         return string
     }
 }
@@ -85,25 +99,25 @@ extension TupleElement {
 extension Type {
     override var description: String {
         var string = "\(type(of: self)): "
-        string += "typealiases = \(typealiases), "
-        string += "isExtension = \(isExtension), "
-        string += "kind = \(kind), "
-        string += "accessLevel = \(accessLevel), "
-        string += "name = \(name), "
-        string += "isGeneric = \(isGeneric), "
-        string += "localName = \(localName), "
-        string += "variables = \(variables), "
-        string += "methods = \(methods), "
-        string += "initializers = \(initializers), "
-        string += "annotations = \(annotations), "
-        string += "staticVariables = \(staticVariables), "
-        string += "instanceVariables = \(instanceVariables), "
-        string += "computedVariables = \(computedVariables), "
-        string += "storedVariables = \(storedVariables), "
-        string += "inheritedTypes = \(inheritedTypes), "
-        string += "containedTypes = \(containedTypes), "
-        string += "parentName = \(parentName), "
-        string += "attributes = \(attributes)"
+        string += "typealiases = \(self.typealiases), "
+        string += "isExtension = \(self.isExtension), "
+        string += "kind = \(self.kind), "
+        string += "accessLevel = \(self.accessLevel), "
+        string += "name = \(self.name), "
+        string += "isGeneric = \(self.isGeneric), "
+        string += "localName = \(self.localName), "
+        string += "variables = \(self.variables), "
+        string += "methods = \(self.methods), "
+        string += "initializers = \(self.initializers), "
+        string += "annotations = \(self.annotations), "
+        string += "staticVariables = \(self.staticVariables), "
+        string += "instanceVariables = \(self.instanceVariables), "
+        string += "computedVariables = \(self.computedVariables), "
+        string += "storedVariables = \(self.storedVariables), "
+        string += "inheritedTypes = \(self.inheritedTypes), "
+        string += "containedTypes = \(self.containedTypes), "
+        string += "parentName = \(self.parentName), "
+        string += "attributes = \(self.attributes)"
         return string
     }
 }
@@ -111,10 +125,10 @@ extension Type {
 extension Typealias {
     override var description: String {
         var string = "\(type(of: self)): "
-        string += "aliasName = \(aliasName), "
-        string += "typeName = \(typeName), "
-        string += "parentName = \(parentName), "
-        string += "name = \(name)"
+        string += "aliasName = \(self.aliasName), "
+        string += "typeName = \(self.typeName), "
+        string += "parentName = \(self.parentName), "
+        string += "name = \(self.name)"
         return string
     }
 }
@@ -122,14 +136,14 @@ extension Typealias {
 extension Variable {
     override var description: String {
         var string = "\(type(of: self)): "
-        string += "name = \(name), "
-        string += "typeName = \(typeName), "
-        string += "isComputed = \(isComputed), "
-        string += "isStatic = \(isStatic), "
-        string += "readAccess = \(readAccess), "
-        string += "writeAccess = \(writeAccess), "
-        string += "annotations = \(annotations), "
-        string += "attributes = \(attributes)"
+        string += "name = \(self.name), "
+        string += "typeName = \(self.typeName), "
+        string += "isComputed = \(self.isComputed), "
+        string += "isStatic = \(self.isStatic), "
+        string += "readAccess = \(self.readAccess), "
+        string += "writeAccess = \(self.writeAccess), "
+        string += "annotations = \(self.annotations), "
+        string += "attributes = \(self.attributes)"
         return string
     }
 }