+++ Sourcery/Models/Class.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-// sourcery: skipDescription, skipEquatable
+// sourcery: skipDescription
 final class Class: Type {
     override var kind: String { return "class" }