+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Sourcery CHANGELOG
 
 ---
+
+## Master
+
+### Bug Fixes
+
+- fixed preserving order of inherited types
+
+### Internal changes
+
+- adjusted internal templates and updated generated code 
+
 ## 0.5.3
 
 ### New Features